Adicionando função updateAutor

+++ src/main/java/com/wanderley/victor/gerenciadorlivrosapi/repositories/AutorRepositoryImpl.java
@@ -7,6 +7,7 @@ package com.wanderley.victor.gerenciadorlivrosapi.repositories;
 
 import com.wanderley.victor.gerenciadorlivrosapi.model.Autor;
 import java.util.List;
+import java.util.Set;
 import org.springframework.stereotype.Repository;
 
 /**
@@ -24,12 +25,27 @@ public class AutorRepositoryImpl implements AutorRepository {
     //<------------------------Métodos insert---------------------------------->
     @Override
     public Boolean addAutor(final Autor autor) {
+        
         return autorDAO.addAutor(autor);
     }
     //<--------------------------Metodos Delete-------------------------------->
     @Override
-    public Boolean deleteAutor(Integer idAutor) {
-        return autorDAO.deleteAutor(idAutor);
+    public Boolean deleteAutor(final Integer idAutor) {
+        if(getIdsValidos().contains(idAutor)){
+            return autorDAO.deleteAutor(idAutor);
+        }return false;
+    }
+    //<--------------------------Métodos Update-------------------------------->
+    @Override
+    public Boolean updateAutor(final Integer id, final Autor autor){
+        if(getIdsValidos().contains(id)){
+            return autorDAO.updateAutor(id, autor);
+        }return false;
+        
+    }
+    
+    private Set<Integer> getIdsValidos(){
+        return autorDAO.idsValidos();
     }
